updating post api functions to check certain things wrt logged in user id

+++ app/routers/post.py
@@ -1,15 +1,22 @@
-from typing import Annotated
+from typing import Annotated, Optional
 from fastapi import APIRouter, HTTPException, Query, status
-from ..models import Post, PostPublic, PostCreate, PostUpdate
+from fastapi.params import Depends
+from ..models import Post, PostPublic, PostBase, PostCreate, PostUpdate, User
+from ..oauth2 import get_current_user
 from ..database import SessionDep
 from sqlmodel import Session, select
 
 router = APIRouter(prefix="/posts", tags=["Post"])
 
 @router.post("/", response_model=PostPublic, status_code = status.HTTP_201_CREATED)
-def create_post(post: PostCreate, session: SessionDep):
+def create_post(session: SessionDep, current_user: Annotated[User, Depends(get_current_user)], post: PostCreate, ):
     try:
-        db_post = Post.model_validate(post)
+        print("input post", type(post), post)
+        db_post = Post(
+                **post.model_dump(),
+                owner_id=current_user.id,
+            )
+        # db_post = Post.model_validate(post)
         session.add(db_post)
         session.commit()
         session.refresh(db_post)
@@ -21,11 +28,16 @@ def create_post(post: PostCreate, session: SessionDep):
 @router.get("/", response_model=list[PostPublic])
 def read_posts(
     session: SessionDep,
+    current_user: Annotated[User, Depends(get_current_user)],
     offset: int = 0,
     limit: Annotated[int, Query(le=100)] = 3,
+    show_all: Optional[bool] = True
 ):
-    posts = session.exec(select(Post).offset(offset).limit(limit)).all()
 
+    select_stmt = select(Post)
+    select_stmt = select_stmt if show_all else select_stmt.where(Post.owner_id == current_user.id)
+    posts = session.exec(select_stmt.offset(offset).limit(limit)).all()
+    
     if not posts:
         raise HTTPException(
             status_code = status.HTTP_404_NOT_FOUND,
@@ -62,14 +74,20 @@ def update_post(post_id: int, post: PostUpdate, session: SessionDep):
     return post_db
 
 
-@router.delete("/{post_id}")
-def delete_post(post_id: int, session: SessionDep):
+@router.delete("/{post_id}", status_code = status.HTTP_204_NO_CONTENT)
+def delete_post(post_id: int, session: SessionDep, current_user: Annotated[User, Depends(get_current_user)],):
     post = session.get(Post, post_id)
     if not post:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, 
             detail= f"Post with id: {post_id} was not found."
         )
+
+    if post.owner_id!=current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN, 
+            detail= "You are not allowed to delete this post."
+        )
+     
     session.delete(post)
     session.commit()
-    return {"ok": True}
